+++ lib/src/models/view_configurations/multi_day_configurations/multi_week_configuration.dart
@@ -10,7 +10,7 @@ class MultiWeekConfiguration extends MultiDayViewConfiguration {
     super.daySeparatorLeftOffset = 8,
     super.hourLineLeftMargin = 56,
     super.multiDayTileHeight = 24,
-    super.paintWeekNumber = true,
+    super.showWeekNumber = true,
     super.eventSnapping = false,
     super.timeIndicatorSnapping = false,
     super.createEvents = true,