+++ lib/src/models/view_configurations/multi_day_configurations/week_configuration.dart
@@ -9,7 +9,7 @@ class WeekConfiguration extends MultiDayViewConfiguration {
     super.daySeparatorLeftOffset = 8,
     super.hourLineLeftMargin = 56,
     super.multiDayTileHeight = 24,
-    super.paintWeekNumber = true,
+    super.showWeekNumber = true,
     super.eventSnapping = false,
     super.timeIndicatorSnapping = false,
     super.createEvents = true,